requirements.txt
requests
google-cloud-bigquery